backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
numpy>=2.1.0
pandas>=2.2.3
python-dotenv==1.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx>=0.25.2
sqlalchemy>=2.0.23
alembic>=1.13.0
psycopg2-binary>=2.9.10
scikit-learn>=1.5.0
joblib>=1.3.0
matplotlib>=3.8.2
seaborn>=0.13.0
geopandas>=0.14.1
shapely>=2.0.2
folium>=0.15.0
requests>=2.31.0
aiofiles>=23.2.1
Pillow>=10.1.0
